Allow child classes to also get selected for bounty targets
There should be an easier way to do this that doesn't involve a second forloop, but for now just checking if a is a child of actorName as well as being explictly the same className should suffice.  This should allow for addons that make enemy variations also be selected for bounties.