Initial Commit to add ability to earn money from normal mobs
Currently adds a "Neither" money mode as well as a non-bounty drop ratio that multiplies the existing bounty amount to get the normal drop rate, which is then used as the upper bound for the random amount a given mob will drop when they die.
MercBucks are now also multi-pickups, which should help with collecting the stacks that some mobs might drop.