Add Magazine Store, Refactor out all addon store entries
These will now live, in some form, in their respective packs.  The MENUDEF still needs to be redone, so for now it'll house the filter toggles same as it did.  In the future I'd prefer it if, like the Spawn Tables, can generate menus based off the commands and their params being CVAR-driven
MercBucks had to be marked as the default currency, as loading Ace-Corp-Extended first caused AssemblyCores to be the default, which is undesired.
Merchant Skins also now get a tag property to further customize their L&F