Add ability for Merchant Stores to use Weighted Random Tables
Now some stores can be more common than others.  By default all stores get a relative weight of 1, so they're just as equally weighted as they were before this change.

Due to the need to pass functions into other functions, this does bump the minimum ZScript version to 4.14, but that's already a bit of a given with HDCoreLib.